requirements.txt
bcrypt==4.0.1
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nulltype==2.3.1
plaid-python==11.6.0
protobuf==3.20.3
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
yelpapi==2.5.0
